+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added adapters for http version 0.2 types.
+- Added `sha256` function to `apx_core::hashes` module.
 
 ### Changed
 
@@ -16,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Re-export `http` types from `http_types` module.
 - Migrated to http package version 1.1.0.
 - Migrated to reqwest version 0.12 and rustls version 0.22.
+- Use `ContentDigest` type during HTTP signature verification.
 
 ## [0.5.0] - 2024-11-20